refactor(presenters): reduce site-to-email coupling by moving shared helpers to shared/
Moved cross-presenter utilities from email/ to shared/ to eliminate
site presenter importing from email presenter:

- Created src/presenters/shared/window-helpers.ts with:
  - displaySessionName, bestTimeLabel, bestDaySessionLabel
  - localSummaryLines, timeAwareLocalSummary
  - isAstroWindow, moonDescriptor, peakHourForWindow
  - displayTag, displayBestTags

- Created src/presenters/shared/kit-advisory.ts with:
  - buildKitTips, evaluateKitRules, KitTip type

- Updated email/time-aware.ts to re-export from shared for backwards compatibility
- Updated email/kit-advisory.ts to re-export from shared for backwards compatibility
- Updated site/format-site.ts to import from shared instead of email
- Updated all relevant READMEs to document the new module structure

Fixes #127
